refactor: modules for worker
