+++ worker/tests/cassandra_client.py
